// src/enums/UPGRADES.ts
export enum UPGRADES {
    SWORD_1 = 0,
    SWORD_2 = 1,
    WEAPON_1 = 2,
    WEAPON = 3,
    ARROW_1 = 4,
    ARROW_2 = 5,
    THROWING_AXES_1 = 6,
    THROWING_AXES_2 = 7,
    HUMAN_SHIELD_1 = 8,
    HUMAN_SHIELD_2 = 9,
    ORC_SHIELD_1 = 10,
    ORC_SHIELD_2 = 11,
    HUMAN_BOAT_ATTACK_1 = 12,
    HUMAN_BOAT_ATTACK_2 = 13,
    ORC_BOAT_ATTACK1 = 14,
    ORC_BOAT_ATTACK2 = 15,
    HUMAN_BOAT_ARMOR_1 = 16,
    HUMAN_BOAT_ARMOR_2 = 17,
    ORC_BOAT_ARMOR_1 = 18,
    ORC_BOAT_ARMOR_2 = 19,
    CATAPULT_DAMAGE_1 = 20,
    CATAPULT_DAMAGE_2 = 21,
    BALLISTA_DAMAGE_1 = 22,
    BALLISTA_DAMAGE_2 = 23,
    TRAIN_RANGERS = 24,
    LONGBOWS = 25,
    RANGER_SCOUTING = 26,
    RANGER_MARKSMANSHIP = 27,
    TRAIN_BERSERKERS = 28,
    LIGHTER_AXES = 29,
    BERSERKER_SCOUTING = 30,
    BERSERKER_REGENERATION = 31,
    TRAIN_OGRE_MAGES = 32,
    TRAIN_PALADINS = 33,
    SPELL_HOLY_VISION = 34,
    SPELL_HEAL = 35,
    SPELL_EXORCISM = 36,
    SPELL_FIRE_SHIELD = 37,
    SPELL_FIREBALL = 38,
    SPELL_SLOW = 39,
    SPELL_INVISIBILITY = 40,
    SPELL_POLYMORPH = 41,
    SPELL_BLIZZARD = 42,
    SPELL_EYE_OF_KILROGG = 43,
    SPELL_BLOODLUST = 44,
    SPELL_RAISE_DEAD = 45,
    SPELL_DEATH_COIL = 46,
    SPELL_WHIRL_WIND = 47,
    SPELL_HASTE = 48,
    SPELL_UNHOLY_ARMOR = 49,
    SPELL_RUNES = 50,
    SPELL_ROT = 51,
}